Created
Created a FeatureSet abstraction that allows us to efficiently extract
and load features for use in a model. The FeatureSet class basically
checks to see if features have already been extracted and saved to
a file on disk. If they have, then simply load them from the file. For
more details on how this works, please refer to
features/example_feature_set.py file.

Turned our project into a python module by adding __init__.py to most
of the directories. This is useful for importing python files in
different directories. For example, to test a model that uses something
from the features directory, you can simply run "python -m models/myModel"
from the root directory of the project. By executing scripts as a module with
the -m flag, you are basically making python aware of all the
submodules in the project.

Retrofitted the naive_bayes model to take advantage of the
ngram_feature_set. Feature extraction is now much more efficient, and I
can actually run a Naive Bayes model on my machine. Hey, we are getting
72% accuracy with Naive Bayes using cross validation! Hurray!

Added lda to the environment.yml and updated the version of
scikit-learn. Please remember to update your environments!