Added the final model with supporting classes
Added the RAOPFieldExtractor class so that we can pass the entire
json data set into the classifier pipeline. The RAOPFieldExtractor
will then pull out the fields we specify and pass those fields to the
next step in the pipeline.

Added the RAOPNumericalFieldExtractor class to pull in all the numerical data
as a feature.

Added the model_runner.py file to run models using a RandomizedSearchCV
class from sklearn. This allows us to randomly try different parameter
sets and then sort the resulting models by their accuracy or some other
metric.

Added the best_tree.py file. This specifies the feature and parameters
for the RandomForestClassifier that we've chosen. This is where you
would add new features to the model

Added the LengthTransformer class to simply extract the length of each
request.