Added Equalized Datasets
I've added a subset of the training data that has an equal number of
positive and negative labels. This will ensure two things:
1) we will have comparable results for our models
2) our models will be more generalized and perform better on the Kaggle
test set.

There are two versions of the JSON format: one format uses ascii-only
characters, while the other one does not. I'm not sure which format will
work for everyone, so if one of them doesn't work, try the other. If
neither work, the CSV format should work because it only contains the
indexes of the records from the original train.json.